Fix reviewing comments
